feat: implement mobile-responsive admin sidebar and enhance product display with list-to-bullet formatting and improved currency input handling

+++ src/lib/validations.ts
@@ -28,7 +28,7 @@ export const productSchema = z.object({
   fallbackGradient: z.string().optional(),
   featured: z.boolean(),
   active: z.boolean(),
-  sortOrder: z.number().int(),
+  sortOrder: z.number(),
   categoryIds: z.array(z.string().uuid()),
 });
